### 2025-04-02 #7982 @editorialbot
[PRE REVIEW]: auto-AUTO: A Python Layer for Automatically Running the AUTO-07p Continuation Software
**Submitting author:** @jodemaey (<a href="http://orcid.org/0000-0002-5098-404X">Jonathan Demaeyer</a>)
**Repository:** https://github.com/Climdyn/auto-AUTO
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5
