### 2025-04-14 36bd2f4
Added a limit parameter for the period of the continuation solutions
