### 2025-04-14 #7982 @epspebble
> 👋  @epspebble, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> In either case, would you be able to recommend other possible reviewers?

Yes, I would be happy to help.
