### 2025-04-15 09a743c
Taking care of the specified parameter name for the period in the code introduced in the previous commits
