### 2025-04-17 #7982 @cdiddens
> 👋 [@cdiddens](https://github.com/cdiddens) & [@duarterocha](https://github.com/duarterocha), would one of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Dear @danielskatz ,

thank you for the kind invitation. I am also quite busy with other reviewer tasks at the moment. Also, my knowledge in AUTO is very limited, so it would be quite some work for me to judge whether the software, which appears to be some Python wrapper for AUTO, adds a lot of value.

I therefore unfortunately have to decline the invitation.

I hope you can find suitable other reviewers. I know that researchers at the Institute of Theoretical Physics in Münster (groups Prof. Thiele & Prof. Gurevich) are using AUTO (and pde2path), but I unfortunately do not know whether they are around here at GitHub and/or willing to review. Beyond that, I cannot recommend any other reviewers. Sorry.
