### 2025-04-17 #7982 @tcriess
Hi @danielskatz - I am really sorry to decline this invitation, but I have not used AUTO (or in fact any other continuation software) for a very long time and don't think that I can really help here... I assume you have invited @bartoldeman (I hope that is the correct github handle)?
Best wishes!
