### 2025-04-17 #7982 @bartoldeman
> 👋 [@bartoldeman](https://github.com/bartoldeman), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sure, this looks like something I can review.
