### 2025-04-18 #7982 @rveltz
Hi, sorry for coming late to the show. I won't be able to review the proposal, I am too busy
