### 2025-04-18 #8079 @danielskatz
👋 @jodemaey - Please add the missing DOI that editorialbot [found](https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2815425414) in your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
