### 2025-04-21 #8079 @danielskatz
Thanks @lockwo - as you can see [above](https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2815428462), your first action is to use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
