### 2025-04-24 #8079 @danielskatz
👋  @bartoldeman - When you get a chance, please use the command `@editorialbot generate my checklist` to create your review checklist.
