+++ documentation/notes/auto2.bib
@@ -15,7 +15,8 @@ hu@article{armstrongmckay2022,
     pages = {1166--1184},
     title = {Tipping Points in Open Systems: Bifurcation, Noise-Induced and Rate-Dependent Examples in the Climate System},
     volume = {370},
-    year = {2012}
+    year = {2012},
+    doi = {10.1098/rsta.2011.0306},
 }
 
 @book{doedel2012,
