### 2025-04-25 0f479f6
Setting a maximum period for periodic orbit continuation (#23)
Added a limit parameter for the period of the continuation solutions
