+++ auto2/diagrams/bifurcations.py
@@ -377,7 +377,7 @@ class BifurcationDiagram(object):
 
     def compute_periodic_orbits_diagram(self, end_level=10, extra_comparison_parameters=None, comparison_tol=2.e-2,
                                         remove_dubious_bp=True, max_number_bp=None, max_number_bp_detected=None,
-                                        backward_bp_continuation=False, **continuation_kwargs):
+                                        backward_bp_continuation=False, maximum_period=None, **continuation_kwargs):
         """Method which starts the computation of the periodic orbits of a bifurcation diagram where
         the fixed points have already been continued. It will start by continuing the Hopf bifurcations, then the resulting branching
         and period doubling points. Each computation is done level by level in the the developing tree of the bifurcation
@@ -416,6 +416,10 @@ class BifurcationDiagram(object):
         backward_bp_continuation: bool, optional
             Compute also the `backward` continuation at the branching points.
             Defaults to `False`.
+        maximum_period: float or None, optional
+            Define a maximum period for the continuation of periodic orbits.
+            If this maximum period is reached for a given solution, the continuations will stop at the previous solution.
+            Disabled if `None`. Default to `None`.
         continuation_kwargs: dict
             Parameters to pass to AUTO for each continuation.
             See the :meth:`~auto2.continuations.periodic_orbits.PeriodicOrbitContinuation.make_continuation` method of the
@@ -483,7 +487,7 @@ class BifurcationDiagram(object):
                     hp.make_continuation(hb, max_bp=max_number_bp_detected, **used_continuation_kwargs)
 
                     logger.debug('Continuation done. Checking now against previous continuation...')
-                    self._check_po_continuation_against_itself(hp, used_continuation_kwargs, extra_comparison_parameters, comparison_tol, max_number_bp_detected)
+                    self._check_po_continuation_against_itself(hp, used_continuation_kwargs, extra_comparison_parameters, comparison_tol, max_number_bp_detected, maximum_period)
 
                     self._check_po_continuation_against_other_fp_branches(hp, used_continuation_kwargs, extra_comparison_parameters, comparison_tol, max_number_bp_detected)
                     valid_branch = self._check_po_continuation_against_other_po_branches(hp, used_continuation_kwargs, extra_comparison_parameters, comparison_tol, max_number_bp_detected)
@@ -517,7 +521,7 @@ class BifurcationDiagram(object):
 
     def restart_periodic_orbits_diagram(self, end_level=10, extra_comparison_parameters=None, comparison_tol=2.e-2,
                                         remove_dubious_bp=True, max_number_bp=None, max_number_bp_detected=None,
-                                        backward_bp_continuation=False, restart=True, **continuation_kwargs):
+                                        backward_bp_continuation=False, maximum_period=None, restart=True, **continuation_kwargs):
         """Method which restarts the computation of the periodic orbits of a bifurcation diagram stopped previously at a certain level
         of the developing tree of the bifurcation diagram.
         Each computation is done level by level in the the developing tree of the bifurcation diagram.
@@ -556,6 +560,10 @@ class BifurcationDiagram(object):
         backward_bp_continuation: bool, optional
             Compute also the `backward` continuation at the branching points.
             Defaults to `False`.
+        maximum_period: float or None, optional
+            Define a maximum period for the continuation of periodic orbits.
+            If this maximum period is reached for a given solution, the continuations will stop at the previous solution.
+            Disabled if `None`. Default to `None`.
         restart: bool, optional
             Whether to restart or simply start the computation of the periodic orbits bifurcation diagram.
             Defaults to `True` which means restart.
@@ -685,7 +693,7 @@ class BifurcationDiagram(object):
[…]
