### 2025-04-25 1cdc3c4
Specifies default for ICP param in case not provided by the user In particular, if period param is missing, add it
