### 2025-05-06 #8079 @lockwo
Review notes:
> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

Although the results are recreate-able by exactly running the example, if the example changes this would not be the case, so maybe the raw data should be there as well

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The core functions are documented, though the “User guide” is pending. I would also recommend embedding the example notebooks into the docs.

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

This is available at the bottom of the readme, but I would recommend and explicit CONTRIBUTING.md

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

As far as I can tell, this is the biggest missing element. There don’t seem to be tests added yet. There are many types tests (from testing correctness, to testing speed, to just making sure the code runs without errors) that would be helpful to have.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

Although similar packages are presented, there isn’t a discussion of how this package is related to them (or the relative features/strengths/weaknesses).

General points:
- The figures in the paper could be higher DPI, usually doing `plt.savefig(, dpi=300)` is good enough
- The codebase would probably benefit from a formatter (e.g. black/ruff) to help increase its readability
