### 2025-05-09 04ce0f1
Fixed a bug in the plot branch routine (#26)
