+++ pyproject.toml
@@ -7,7 +7,7 @@ requires = ["setuptools", "wheel"]
 [project]
 requires-python = ">=3.10"
 name = "auto-AUTO"
-version = "0.5.2"
+version = "0.5.3"
 description = "An AUTO-07p automatic search algorithm codebase"
 readme = "README.md"
 authors = [
