+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - numba
   - defaults
 dependencies:
-  - python>=3.8
+  - python>=3.10
   - ipython
   - jupyter
   - matplotlib>=3.4
