### 2025-05-11 c1c8056
Implemented the plot of PO branches with 3D solutions plot
