+++ auto2/continuations/base.py
@@ -1492,11 +1492,11 @@ class Continuation(ABC):
             Size of the text plotted to display the bifurcation points. Defaults to `12`.
         marker : str, optional
             The marker style used for plotting fixed points. If `None`, set to default marker.
-        linestyle : str, optional
+        linestyle: str, optional
             The line style used for plotting the periodic orbit solutions. If `None`, set to default marker.
-        linewidth : float, optional
+        linewidth: float, optional
             The width of the lines showing the periodic solutions. If `None`, set to the default.
-        color_solutions : bool, optional
+        color_solutions: bool, optional
             Whether to color each solution differently. In this case, a valid `parameter` argument must be provided.
             If `True`, and `parameter` argument is provided and valid, then solutions are colored based on the parameter value of a given solution.
             Use the cmap provided in the `plot_kwargs` argument. If no cmap is provided there, use `Blues` cmap by default.
@@ -1627,7 +1627,7 @@ class Continuation(ABC):
         return ax
 
     def plot_solutions_3D(self, variables=(0, 1, 2), ax=None, figsize=(10, 8), markersize=None, marker=None, linestyle=None,
-                       linewidth=None, plot_kwargs=None, labels=None, indices=None, parameter=None, value=None,
+                       linewidth=None, plot_kwargs=None, labels=None, color_solutions=False, indices=None, parameter=None, value=None,
                        variables_name=None, tol=0.01):
         """
         Plots the stored solutions for a given set of variables, using a 3-dimensional projection.
@@ -1653,10 +1653,16 @@ class Continuation(ABC):
             Size of the text plotted to display the bifurcation points. Defaults to `12`.
         marker : str, optional
             The marker style used for plotting fixed points. If `None`, set to default marker.
-        linestyle : str, optional
+        linestyle: str, optional
             The line style used for plotting the periodic orbit solutions. If `None`, set to default marker.
-        linewidth : float, optional
+        linewidth: float, optional
             The width of the lines showing the periodic solutions. If `None`, set to the default.
+        color_solutions: bool, optional
+            Whether to color each solution differently. In this case, a valid `parameter` argument must be provided.
+            If `True`, and `parameter` argument is provided and valid, then solutions are colored based on the parameter value of a given solution.
+            Use the cmap provided in the `plot_kwargs` argument. If no cmap is provided there, use `Blues` cmap by default.
+            If `False`, all solutions are colored identically, controlled using `plot_kwargs`.
+            Default is `False`.
         plot_kwargs: dict or None, optional
             Additional keyword arguments for the plotting function. Default is `None`.
         labels: str or list(str), optional
@@ -1716,38 +1722,62 @@ class Continuation(ABC):
         if plot_kwargs is None:
             plot_kwargs = dict()
 
+        # Colouring solutions dependant on parameter value
+        if 'cmap' in plot_kwargs:
+            cmap = plot_kwargs['cmap']
+            if isinstance(cmap, str):
+                cmap = plt.get_cmap(cmap)
+            plot_kwargs.pop('cmap')
+        else:
+            cmap = plt.get_cmap('Blues')
+        if color_solutions:
+            if parameter is not None:
+                if labels is None:
+                    solutions_types = 'all'
+                else:
+                    solutions_types = labels
+                p_vals = self.solutions_parameters(parameters=parameter, solutions_types=solutions_types)
+                p_min, p_max = np.min(p_vals), np.max(p_vals)
+
+                if value is None:
+                    value = p_vals
+            else:
+                raise ValueError('If `color_solutions` argument is set to true, you must provide the `parameter` argument as well.')
+
[…]
