### 2025-05-13 #8079 @danielskatz
👋 @lockwo & @jodemaey - It's good to see your discussion above. As there are many points there, please consider if it would be useful to create any issues in the source repo to separate and track the different points.  If so, please do mention `openjournals/joss-reviews#8079` in such issues so that a link is created to this thread
