### 2025-05-13 #8079 @lockwo
> I don't understand this, there is no raw data provided to the code. What is provided to AUTO (and therefore to auto-AUTO) are equations files. Please explain what raw data you are talking about.

Yes my point was unclear (but thinking about it more, also not that important), all is good in this regard.

> Indeed, the user guide is pending. I hope to have the time to do it before next September. As for the example notebooks in the docs, I did that for a previous project and my experience is that it makes the repo heavier, due to the possible archiving of the past notebooks version figures. This time I wanted to avoid this, but fair enough. I will investigate a solution for this.

That's an understandable goal, I wouldn't want bloat in the repo. I was using sphinx recently and had some success with making a /examples/ in docs that is just a symlink to the examples main directory. Then with `nbsphinx` and `nbsphinx_link` (and setting to never execute the files) in the conf.py. But I am not knowledgeable enough of sphinx to know if this is universally applicable.

> This is indeed missing. My current idea of testing this was to pytest the notebooks. This could be done soon, during the present review.
>
> We could make more atomic tests, but then it would require some work of identifying the dynamical systems which present corresponding features to be tested (like continuation branches merging, crossing or looping over themselves). I (or [@ushham](https://github.com/ushham)) can do that in the future, but right now we don't have the bandwidth.

I understand these full feature atomic tests could be quite some difficulty, soI think the pytest of the notebooks is definitely good. That ensures everything is working as expected.

> It is true that there is no real comparison with other continuation packages, because that would be tantamount to comparing AUTO and the others. auto-AUTO on the other hand, is a layer on top of AUTO, implementing some data management, automatisms and continuation rules, and also plotting features, all of this in notebooks (while the underlying AUTO software has its own Python CLI). As such, there is not exactly an equivalent to auto-AUTO, as it is not per se a continuation software.

I agree with this point, that auto^2 stands somewhat apart from other packages, and the features are not 1 to 1 comparable. That being said, perhaps it is worth including brief remark to this effect in the comparison section? I haven't worked much with the packages listed there, so I leave this to your discretion.

> My personal view on this is that I use formatter for operational codes only. However, I would consider this if the number of contributors increases dramatically, but otherwise not.

Fair enough, I suppose my proclivity for black is more of a philosophical opinion. After all, PEP 8 is really a [guideline more than a rule](https://www.youtube.com/watch?v=k9ojK9Q_ARE) 😄

> 👋 [@lockwo](https://github.com/lockwo) & [@jodemaey](https://github.com/jodemaey) - It's good to see your discussion above. As there are many points there, please consider if it would be useful to create any issues in the source repo to separate and track the different points. If so, please do mention `openjournals/joss-reviews#8079` in such issues so that a link is created to this thread

Sure, I just made one for the contributing.md, that's a small fix but easy to track via issues rather than through multiple messages.
