### 2025-05-13 Climdyn/auto-AUTO#28 @lockwo
Add contributing.md
As discussed in the review https://github.com/openjournals/joss-reviews/issues/8079, having an explicit contributing.md allows GitHub to parse it and can be beneficial.
