### 2025-05-13 #8079 @epspebble
@danielskatz I am just freed to start reviewing yesterday. Currently I agree with other reviewers that the documentation is rather sparse. For the two examples that I can find, in the form of a jupyter notebook, they are only readable to people well versed in AUTO-07p.

The overall premise about the necessity of this software for systems in high dimensional continuation problems is legitimate in my opinion, but I fail to find any information on how it is applied to a >10 parameter system.

I would try to put in some suggestions in the form of issues as you say, but it would be nice to refer us to some examples of them. The work that can be done could be rather substantial.
