### 2025-05-14 #8079 @danielskatz
@epspebble - I'm not sure if I completely understand your last point.

Re issues, I think this could be as simple as creating a new issue in https://github.com/Climdyn/auto-AUTO/issues/ saying "More documentation is needed" and then you and the other reviewers and the authors could discuss what is (minimally) needed for the JOSS review to be complete, what would be practical to add now for the review, and what should be added in the future but isn't needed for the JOSS review to be successful.
