### 2025-05-15 Climdyn/auto-AUTO#29 @epspebble
Further work to complete the enz example.
At the end of the notebook, it says "You can now explore continuations in other parameters:"

Could you show how you can restart computations at one of the branch points found but continue in another parameter, say s2?

openjournals/joss-reviews#8079
