### 2025-05-15 Climdyn/auto-AUTO#30 @epspebble
One more example each for fixed_points and periodic_orbits
Although enz and lrz are great choices from the AUTO-07p demos for, respectively, fixed points and periodic orbits calculations, I do not have enough clarity how a user could use your examples towards their own calculations.  One more example each may help:

For fixed point calculations, I suggest dd2.
For periodic orbit calculations, I suggest plp.

Then, since the main impetus for this software concerns systems with high number of variables, could you give a demo (maybe simplified version suffices) for your study of the qgs model?

openjournals/joss-reviews#8079
