### 2025-05-16 #8079 @jodemaey
> [@danielskatz](https://github.com/danielskatz) I am just freed to start reviewing yesterday. Currently I agree with other reviewers that the documentation is rather sparse. For the two examples that I can find, in the form of a jupyter notebook, they are only readable to people well versed in AUTO-07p.

Hi @epspebble , first thanks a lot for your time. Then, indeed, it is assumed from the start that people need to look into AUTO documentation and familiarized themselves with it first before using auto-AUTO. What would be super nice of course is to really provide a tutorial which teach people about continuation and AUTO at the same time, using notebooks. This is because learning AUTO is a long and steep process, involving also possibly learning dynamical system theory. AUTO does that through the examples, but there was always a gap that the users have to fill by themselves. So this would be nice but would require a lot of work and time (teaching properly is one of the hardest thing imho) that we don't have for the moment. Somehow this is a project in itself, which we could think about (by we I mean all interested in this) in the future.

But I think improving and providing more examples is the way to go for the present review. I will address this (via the issues you raised).

I will also comment on the high-dimensional cases later today (I have to go now).
