### 2025-05-16 #8079 @jodemaey
Hi @lockwo ,

Just to let you know that finally, I changed my mind, and I think I will enforce black soon (you can create an issue about that if you want).
Thanks for the suggestion.
