### 2025-05-17 d63e95b
Adding a CONTRIBUTING file
