+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+
+Contributing to auto-AUTO
+=========================
+
+If you want to contribute actively, please contact the main authors.
+
+In addition, if you have made changes that you think will be useful to others, please feel free to suggest these
+as a pull request on the [auto-AUTO Github repository](https://github.com/Climdyn/auto-AUTO).
+
