### 2025-05-19 Climdyn/auto-AUTO#30 @ushham
Thanks you @epspebble, I will add the three suggestions that you suggested in the coming days.
