### 2025-05-19 cce3239
Specifies default for ICP param in case not provided by the user (#24)
* Specifies default for ICP param in case not provided by the user
In particular, if period param is missing, add it.

* Solved a bug with the logic
