### 2025-05-20 Climdyn/auto-AUTO#32 @lockwo
Format code
As per https://github.com/openjournals/joss-reviews/issues/8079, code can be formatted with tools such as black/ruff/etc. These tools can also be setup as GitHub workflows to be checked on PRs, or via pre-commit, or just run by the user.
