### 2025-05-20 #8079 @epspebble
Thank you for committing to work on a comment I gave. I also strongly agree that XPPAUT should be cited. There used to be an iPad app for it too (was it Bart?). Anyways, I applaud your effort to improve the usability of the reliable and trusted AUTO code.
