### 2025-05-22 ee30905
Added a function to return the continuation points parameter values
