### 2025-05-23 Climdyn/auto-AUTO#28 @jodemaey
Closed by #31
