### 2025-05-27 #8079 @jodemaey
Hi @danielskatz ,

So from what I see, here is a checklist of things that still need to be done:

- [ ] Add more example notebooks (from AUTO examples)
- [ ] Make a "CI" pipeline (black, notebook tests)
- [ ] Add a high-dimensional example case
- [ ] Write the Userguide
- [ ] Add more context in the article about auto-AUTO (cite and compare with XPPAuth, and the other libraries proposed by @bartoldeman )
- [ ] Produce higher resolution figures for the article

Please tell me if something is missing, and I will add it.

We can surely manage all of these before mid-July, maybe end of June, but not the Userguide, which I planned to do in August - September. Right now, I cannot work on auto-AUTO until mid-next week. When I resume working on it, I will see how fast I go. Hopefully it will be done faster than what I think.

Will you need a final answer to the referees ?

Thanks,

Jonathan
