### 2025-05-27 #8079 @jodemaey
> The overall premise about the necessity of this software for systems in high dimensional continuation problems is legitimate in my opinion, but I fail to find any information on how it is applied to a >10 parameter system.

Hi again @epspebble ,

So here is my answer to you about high dimensional systems: Indeed auto-AUTO was designed because when dealing with high-dimensional systems, most of the time you don't know where to start, and you need to be systematic. So indeed having a better description about how to deal with high-dimensional system should be in doc (part of the future Userguide maybe). This we will do for sure.

But having practical examples is a bit more complicated, because high-dimensional systems requires a lot of computation power and/or time. For example, getting the full results in https://arxiv.org/pdf/2503.02808 took us weeks of computation overall with a CPU intensive cluster. So it would probably not be a practical example. But on the other hand, if we take a too simple system (coming out of qgs for example, as you suggest in https://github.com/Climdyn/auto-AUTO/issues/30), then it can be "trivial" and you miss the point of using auto-AUTO. AUTO alone is enough then.

So we need to find "perfect" examples: computable with not too much efforts, and complex enough so that there is a need for auto-AUTO.

With @ushham, we have some ideas about how to do it with qgs, but any suggestion about this is welcome (maybe with other models). We will come back with some propositions soon.
