### 2025-07-14 c211dc4
As suggested by Oisin, warning and switching to cmap instead of crashing
