### 2025-07-14 4b8547d
Cleaning and commenting
