### 2025-07-14 0c4ea18
Now checking for the case where there are too many branches to use TABLEAU color list in plots. (#25)
Warning users and switching to a default cmap.
