### 2025-07-14 e0d6cbe
Implemented the plot of PO branches with 3D solutions plot (#35)
