+++ auto2/diagrams/bifurcations.py
@@ -2452,6 +2452,143 @@ class BifurcationDiagram(object):
 
         return axes
 
+    def plot_single_po_branch_and_solutions_3D(self, branch_number, parameter=None, diagram_variables=(1,), solutions_variables=(0, 1, 2),
+                                               axes=None, figsize=(10, 16), solutions_tol=0.01, cmap=None, branch_diagram_kwargs=None,
+                                               solution_diagram_kwargs=None, point_solutions_types=('HB', 'BP', 'UZ', 'PD', 'TR', 'LP')):
+        """Plots the bifurcation diagram of a single branch with the parameter values of stored solutions plotted on this branch,
+        along with a plot of the projection of all solutions.
+
+        Parameters
+        ----------
+        branch_number: int
+            The number of the branch to plot.
+        parameter: str, optional
+            The parameter along which to plot the diagram and solutions.
+        diagram_variables: tuple(int or str, int or str), optional
+            The index or label of the variables along which the bifurcation diagram is plotted.
+            If labels are used, it should be labels of the available monitored variables during the continuations
+            (if their labels have been defined in the AUTO configuration file) given by :meth:`available_variables`.
+            The first value in the tuple determines the variable plot on the x-axis, and the second the y-axis
+            Defaults to `(1, )`.
+        solutions_variables: tuple(int or str, int or str), optional
+            The index or label of the variables shown in the solution plot.
+            If labels are used, it should be labels of the phase space variables
+            (if they have been defined in the AUTO configuration file) given by :meth:`phase_space_variables`.
+            The first value in the tuple determines the variable plot on the x-axis, the second determines the y-axis.
+            Defaults to the first two variables: `(0, 1)`.
+        axes: numpy.ndarray(matplotlib.axes.Axes) or None, optional
+            An array of two |Matplotlib| axis object to plot on. If `None`, a new figure is created.
+        figsize: tuple(float, float), optional
+            Dimension of the figure that is returned, only used if `ax` is not passed.
+            Defaults to `(10, 16)`.
+        solutions_tol:  float or list(float) or ~numpy.ndarray(float), optional
+            The numerical tolerance used when comparing solutions to display with the selected `solutions_parameter_value`.
+            Default to `0.01`.
+        cmap: matplotlib.colors.Colormap or str or None, optional
+            A colormap that controls the color of the branches.
+            The branches are colored from the colormap based on the value of the branch number.
+            If `None`, defaults to `Reds` colormaps.
+        branch_diagram_kwargs: dict or None, optional
+            Additional keyword arguments for the bifurcation diagram plotting function. Default is `None`.
+        solution_diagram_kwargs: dict or None, optional
+            Additional keyword arguments for the solution plotting function. Default is `None`.
+        point_solutions_types: list(str), optional
+            The types of solution to plot as point if `plot_sol_points` is `True`.
+            Default to `['HB', 'BP', 'UZ', 'PD', 'TR', 'LP']`.
+
+
+        The parameters below can be included in `solutions_kwargs` for controlling the solution plots.
+        See :meth:`auto2.continuations.base.Continuation.plot_solutions` for more details.
+
+        Keyword Arguments
+        -----------------
+        markersize: float, optional
+            Size of the text plotted to display the bifurcation points. Defaults to `12`.
+        marker : str, optional
+            The marker style used for plotting fixed points. If `None`, set to default marker.
+        linestyle : str, optional
+            The line style used for plotting the periodic orbit solutions. If `None`, set to default marker.
+        linewidth : float, optional
[…]
