+++ environment.yml
@@ -17,3 +17,4 @@ dependencies:
   - pip
   - pip:
       - sphinxcontrib-bibtex
+      - nbmake
