+++ .github/workflows/notebook-tests.yml
@@ -41,14 +41,14 @@ jobs:
           hashFiles('environment.yml') }}
 
     - uses: conda-incubator/setup-miniconda@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-          miniforge-version: latest
-          use-mamba: true
-          activate-environment: auto2
-          environment-file: environment.yml
-          auto-activate-base: false
-          use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
+      with:
+        python-version: ${{ matrix.python-version }}
+        miniforge-version: latest
+        use-mamba: true
+        activate-environment: auto2
+        environment-file: environment.yml
+        auto-activate-base: false
+        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
 
     - shell: bash -el {0} 
       run: |
