### 2025-07-14 fcb7e03
Avoiding Anaconda default channel
