+++ test/install_auto.sh
@@ -1,3 +1,4 @@
+cd $1
 git clone https://github.com/auto-07p/auto-07p.git
 cd auto-07p
 ./configure
