+++ .github/workflows/notebook-tests.yml
@@ -69,6 +69,6 @@ jobs:
       run: |
         cd ../auto-07p/cmds
         source auto.env.sh
-        cd ../auto-AUTO
+        cd ../../auto-AUTO
         python -m pytest --nbmake "./notebooks"
 
