### 2025-07-14 78b49c4
Blackened the code
