+++ documentation/source/conf.py
@@ -11,15 +11,16 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'auto-AUTO'
-copyright = '2025, Jonathan Demaeyer and Oisín Hamilton'
-author = 'Jonathan Demaeyer and Oisin Hamilton'
-release = 'v0.5.3'
+project = "auto-AUTO"
+copyright = "2025, Jonathan Demaeyer and Oisín Hamilton"
+author = "Jonathan Demaeyer and Oisin Hamilton"
+release = "v0.5.3"
 version = release
 
 # -- General configuration ---------------------------------------------------
@@ -29,22 +30,22 @@ version = release
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.mathjax',
-    'sphinx.ext.ifconfig',
-    'sphinx.ext.viewcode',
-    'sphinx.ext.autodoc',
-    'sphinx.ext.doctest',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.todo',
-    'sphinx.ext.coverage',
+    "sphinx.ext.mathjax",
+    "sphinx.ext.ifconfig",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.doctest",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.todo",
+    "sphinx.ext.coverage",
     #    'nbsphinx',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.autosectionlabel',
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosectionlabel",
     # 'sphinxcontrib.bibtex'
 ]
 
-templates_path = ['_templates']
-exclude_patterns = ['files/notebooks/*', '**.ipynb_checkpoints']
+templates_path = ["_templates"]
+exclude_patterns = ["files/notebooks/*", "**.ipynb_checkpoints"]
 
 # -- Debugging ---------------------------------------------------------------
 
@@ -63,23 +64,24 @@ exclude_patterns = ['files/notebooks/*', '**.ipynb_checkpoints']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
-html_static_path = ['_static']
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
 
 # -- Options for LaTeX output -------------------------------------------------
 
 latex_elements = {
-    'printindex': r'\def\twocolumn[#1]{#1}\printindex',
+    "printindex": r"\def\twocolumn[#1]{#1}\printindex",
 }
 
 
 # -- Extension configuration -------------------------------------------------
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'python': ('https://docs.python.org/3', None),
-                       'numpy': ('https://numpy.org/doc/stable/', None),
-                       'scipy': ('https://docs.scipy.org/doc/scipy/', None),
-                       'matplotlib': ('https://matplotlib.org/stable/', None),
-                       }
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "numpy": ("https://numpy.org/doc/stable/", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/", None),
+    "matplotlib": ("https://matplotlib.org/stable/", None),
+}
 
 # Napoleon settings
 napoleon_google_docstring = True
