+++ test/install_auto.sh
@@ -1,3 +1,6 @@
+#!/bin/bash
+# Script to install AUTO on Github runners
+
 cd $1
 git clone https://github.com/auto-07p/auto-07p.git
 cd auto-07p
