### 2025-07-14 9cf35da
CI tests (#37)
Added test workflows
