+++ environment.yml
@@ -2,7 +2,6 @@ name: auto2
 channels:
   - conda-forge
   - numba
-  - defaults
 dependencies:
   - python>=3.10
   - ipython
@@ -17,3 +16,4 @@ dependencies:
   - pip
   - pip:
       - sphinxcontrib-bibtex
+      - nbmake
