### 2025-07-14 edbf643
Linting of the code
