+++ auto2/diagrams/bifurcations.py
@@ -1,4 +1,3 @@
-
 """
 
 Bifurcation diagram class definition
@@ -21,22 +20,27 @@ from matplotlib.lines import Line2D
 from matplotlib.colors import TABLEAU_COLORS
 
 from auto2.parsers.config import ConfigParser
-logger = logging.getLogger('logger')
-logger.info('Using auto-AUTO (AUTO² or auto2) -- An AUTO-07p automatic search algorithm codebase')
-logger.info('Read AUTO-07p manual first before using it. Wishing you a happy continuation, have fun !')
+
+logger = logging.getLogger("logger")
+logger.info(
+    "Using auto-AUTO (AUTO² or auto2) -- An AUTO-07p automatic search algorithm codebase"
+)
+logger.info(
+    "Read AUTO-07p manual first before using it. Wishing you a happy continuation, have fun !"
+)
 logger.info('Logging messages can be found in the file "auto2.log"')
 
 try:
-    auto_directory = os.environ['AUTO_DIR']
+    auto_directory = os.environ["AUTO_DIR"]
 
     for path in sys.path:
         if auto_directory in path:
             break
     else:
         # sys.path.append(auto_directory + '/python/auto')
-        sys.path.append(os.path.join(auto_directory, 'python'))
+        sys.path.append(os.path.join(auto_directory, "python"))
 except KeyError:
-    logger.warning('Unable to find auto directory environment variable.')
+    logger.warning("Unable to find auto directory environment variable.")
 
 from auto2.continuations.fixed_points import FixedPointContinuation
 from auto2.continuations.periodic_orbits import PeriodicOrbitContinuation
@@ -154,11 +158,15 @@ class BifurcationDiagram(object):
             self.set_path_name(path_name)
 
         if model_name is not None:
-            filepath = 'c.'+model_name if self._path_name is None else os.path.join(self._path_name, 'c.'+model_name)
+            filepath = (
+                "c." + model_name
+                if self._path_name is None
+                else os.path.join(self._path_name, "c." + model_name)
+            )
             self.config_object = ConfigParser(filepath)
         else:
             self.config_object = None
-            
+
         self.fp_branches = dict()
         self.po_branches = dict()
 
@@ -180,11 +188,13 @@ class BifurcationDiagram(object):
 
         self.level_reached = 0
 
-        self._comparison_solutions_types = ('HB', 'BP', 'UZ', 'PD', 'EP', 'TR', 'LP')
+        self._comparison_solutions_types = ("HB", "BP", "UZ", "PD", "EP", "TR", "LP")
         self._figure_legend_handles = list()
         self._figure_3d_legend_handles = list()
 
-        self._default_cmap = 'viridis'  # Should be put in some general config file somewhere later
+        self._default_cmap = (
+            "viridis"  # Should be put in some general config file somewhere later
+        )
 
     @property
     def path_name(self):
@@ -202,10 +212,18 @@ class BifurcationDiagram(object):
         if os.path.exists(path_name):
             self._path_name = path_name
         else:
-            warnings.warn("Path name given does not exist. Using the current working directory.")
+            warnings.warn(
+                "Path name given does not exist. Using the current working directory."
+            )
             self._path_name = None
 
-    def compute_fixed_points_diagram(self, initial_points=None, extra_comparison_parameters=None, comparison_tol=2.e-2, **continuation_kwargs):
+    def compute_fixed_points_diagram(
+        self,
+        initial_points=None,
+        extra_comparison_parameters=None,
+        comparison_tol=2.0e-2,
+        **continuation_kwargs,
+    ):
         """Method which starts the computation of a fixed points bifurcation diagrams using a set of provided initial points.
 
         Parameters
@@ -241,53 +259,86 @@ class BifurcationDiagram(object):
 
         """
 
[…]
