+++ auto2/continuations/fixed_points.py
@@ -1,9 +1,9 @@
 """
 
-    Fixed Point Continuation class definition
-    =========================================
+Fixed Point Continuation class definition
+=========================================
 
-    This module implements the fixed point continuation methods.
+This module implements the fixed point continuation methods.
 
 """
 
@@ -14,19 +14,19 @@ import logging
 import traceback
 import glob
 
-logger = logging.getLogger('logger')
+logger = logging.getLogger("logger")
 
 try:
-    auto_directory = os.environ['AUTO_DIR']
+    auto_directory = os.environ["AUTO_DIR"]
 
     for path in sys.path:
         if auto_directory in path:
             break
     else:
         # sys.path.append(auto_directory + '/python/auto')
-        sys.path.append(auto_directory + '/python')
+        sys.path.append(auto_directory + "/python")
 except KeyError:
-    logger.warning('Unable to find auto directory environment variable.')
+    logger.warning("Unable to find auto directory environment variable.")
 
 import auto.AUTOCommands as ac
 import auto.runAUTO as ra
@@ -73,12 +73,14 @@ class FixedPointContinuation(Continuation):
         Continuation.__init__(self, model_name, config_object, path_name)
 
         # plots default behaviours
-        self._default_marker = 'x'
-        self._default_markersize = 6.
-        self._default_linestyle = ' '
+        self._default_marker = "x"
+        self._default_markersize = 6.0
+        self._default_linestyle = " "
         self._default_linewidth = 1.2
 
-    def make_continuation(self, initial_data, auto_suffix="", only_forward=False, **continuation_kwargs):
+    def make_continuation(
+        self, initial_data, auto_suffix="", only_forward=False, **continuation_kwargs
+    ):
         """Make both forward and backward (if possible) continuation of fixed points.
 
         Parameters
@@ -224,7 +226,9 @@ class FixedPointContinuation(Continuation):
         if auto_suffix:
             self.auto_save(auto_suffix)
 
-    def make_forward_continuation(self, initial_data, auto_suffix="", **continuation_kwargs):
+    def make_forward_continuation(
+        self, initial_data, auto_suffix="", **continuation_kwargs
+    ):
         """Make the forward continuation of fixed points.
 
         Parameters
@@ -248,12 +252,14 @@ class FixedPointContinuation(Continuation):
         runner = ra.runAUTO()
         ac.load(self.model_name, runner=runner)
 
-        if 'MXBF' in continuation_kwargs:
-            warnings.warn('Disabling automatic continuation of branch points (MXBF set to 0)')
-        continuation_kwargs['MXBF'] = 0
+        if "MXBF" in continuation_kwargs:
+            warnings.warn(
+                "Disabling automatic continuation of branch points (MXBF set to 0)"
+            )
+        continuation_kwargs["MXBF"] = 0
 
-        if 'IBR' not in continuation_kwargs and self.branch_number is not None:
-            continuation_kwargs['IBR'] = self.branch_number
+        if "IBR" not in continuation_kwargs and self.branch_number is not None:
+            continuation_kwargs["IBR"] = self.branch_number
 
         self.initial_data = initial_data
 
@@ -263,39 +269,47 @@ class FixedPointContinuation(Continuation):
                     cf = ac.run(initial_data, runner=runner, **continuation_kwargs)
                 except AUTORuntimeError:
                     print(traceback.format_exc())
-                    warnings.warn('AUTO continuation failed, possibly retrying.')
+                    warnings.warn("AUTO continuation failed, possibly retrying.")
                 else:
                     break
             else:
-                warnings.warn('Problem to complete the forward AUTO continuation, returning nothing.')
+                warnings.warn(
+                    "Problem to complete the forward AUTO continuation, returning nothing."
[…]

[…]
