### 2025-07-29 #8079 @jodemaey
Hi @danielskatz ,

Yes I have started working on it and already checked some of the boxes above but now I'm in summer break.
I will resume working on it soon.

I still plan to finish everything and respond to the reviewers by the end of September, so present review status is fine imho.
