### 2025-08-06 Climdyn/auto-AUTO#29 @jodemaey
Hi @epspebble ,

Just to let you know that your request drove me to reflect on the fact that it would need a new functionality that is absent for now: the idea would be able to restart a new bifurcation diagram - along a new parameter - from all the solutions found at a given value of the original continuation parameter.

It would be implemented as a method of the `BifurcationDiagram` object creating a new bifurcation diagram object and returning it once computations (up to a given level) have been done.

Because of that, I will not focus on this for now, I will first complete the other requests of the review (except maybe the user guide and then return to it).

Thanks again for this nice idea.
