### 2025-08-06 Climdyn/auto-AUTO#30 @jodemaey
Hi @epspebble ,

I did not implement dd2 because it is a map, and auto-AUTO was not meant for maps but for flows. Maybe it works but I don't want to test that now.

So if you have another idea that would be great.

On the other hand I have added 3 new demos:

- plp (as suggested, thanks !)
- tor
- tor-shil (a variant of tor)
- pp3

so from my point of view there is enough diversity now.

I will mark that as done in the review, but I will also add high-d examples this week, as you requested.

Can I close this issue?
