### 2025-08-07 3b94228
Started coding the RP1982 high-dim demo
