+++ notebooks/geophysics/RP1982/qgs_RP1982.f90
@@ -1,0 +1,446 @@
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!   AUTO file for qgs model
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+
+SUBROUTINE FUNC(NDIM,U,ICP,PAR,IJAC,F,DFDU,DFDP)
+	!--------- ----
+
+	! Evaluates the algebraic equations or ODE right hand side
+
+	! Input arguments :
+	!      NDIM   :   Dimension of the algebraic or ODE system 
+	!      U      :   State variables
+	!      ICP    :   Array indicating the free parameter(s)
+	!      PAR    :   Equation parameters
+
+	! Values to be returned :
+	!      F      :   Equation or ODE right hand side values
+  
+	! Normally unused Jacobian arguments : IJAC, DFDU, DFDP (see manual)
+
+	IMPLICIT NONE
+	INTEGER, INTENT(IN) :: NDIM, IJAC, ICP(*)
+	DOUBLE PRECISION, INTENT(IN) :: U(NDIM), PAR(*)
+	DOUBLE PRECISION, INTENT(OUT) :: F(NDIM)
+	DOUBLE PRECISION, INTENT(INOUT) :: DFDU(NDIM,NDIM),DFDP(NDIM,*)
+
+	DOUBLE PRECISION thetas_1
+	DOUBLE PRECISION hk_2
+	DOUBLE PRECISION k_d
+	DOUBLE PRECISION k_p
+	DOUBLE PRECISION sigma
+
+	thetas_1 = PAR(1)
+	hk_2 = PAR(2)
+	k_d = PAR(3)
+	k_p = PAR(4)
+	sigma = PAR(5)
+
+
+	F(1) =	 -0.5*U(1)*k_d + 0.5*U(11)*k_d - 0.780274140669492*U(13)*hk_2 + 0.780274140669492&
+		&*U(3)*hk_2
+
+	F(2) =	 -0.980418808722262*U(1)*U(3) - 0.980418808722262*U(11)*U(13) + 0.5*U(12)*k_d - 1&
+		&.56867009395562*U(14)*U(16) - 1.50055762081784*U(15)*U(18) + 1.50055762081784*U(&
+		&16)*U(17) - 0.5*U(2)*k_d + 0.101317695204044*U(3) - 1.56867009395562*U(4)*U(6) -&
+		& 1.50055762081784*U(5)*U(8) + 1.50055762081784*U(6)*U(7)
+
+	F(3) =	 0.980418808722261*U(1)*U(2) - 0.290064736308361*U(1)*hk_2 + 0.980418808722261*U(&
+		&11)*U(12) + 0.290064736308361*U(11)*hk_2 + 0.5*U(13)*k_d + 1.56867009395562*U(14&
+		&)*U(15) + 1.50055762081784*U(15)*U(17) + 1.50055762081784*U(16)*U(18) - 0.101317&
+		&695204044*U(2) - 0.5*U(3)*k_d + 1.56867009395562*U(4)*U(5) + 1.50055762081784*U(&
+		&5)*U(7) + 1.50055762081784*U(6)*U(8)
+
+	F(4) =	 3.74531587521356*U(10)*U(7) + 1.87265793760678*U(12)*U(16) - 1.87265793760678*U(&
[…]
