### 2025-08-07 9eea9a8
Continued the RP1982 demo Still need to be tested
