### 2025-08-07 797e0ab
RP1982 high-dimensional model demo (#42)
