### 2025-08-15 Climdyn/auto-AUTO#30 @epspebble
Hi @jodemaey, apologies I wasn't able to review the new things you put it in - I am at my kids' sport competition event away from my office. I will be able to go through it after this weekend. I hope you don't mind waiting just a bit longer. It does sound like you're done a great deal of improvements! Thank you so much on acting on my comments.
