### 2025-08-18 defbb2d
Done docstring of the new restart and change cont param method
