### 2025-08-18 06f7c93
Coded restart along new param Added bifurcation object based logger WIP
