+++ auto2/diagrams/bifurcations.py
@@ -21,14 +21,7 @@ from matplotlib.colors import TABLEAU_COLORS
 
 from auto2.parsers.config import ConfigParser
 
-logger = logging.getLogger("logger")
-logger.info(
-    "Using auto-AUTO (AUTO² or auto2) -- An AUTO-07p automatic search algorithm codebase"
-)
-logger.info(
-    "Read AUTO-07p manual first before using it. Wishing you a happy continuation, have fun !"
-)
-logger.info('Logging messages can be found in the file "auto2.log"')
+general_logger = logging.getLogger("general_logger")
 
 try:
     auto_directory = os.environ["AUTO_DIR"]
@@ -40,7 +33,7 @@ try:
         # sys.path.append(auto_directory + '/python/auto')
         sys.path.append(os.path.join(auto_directory, "python"))
 except KeyError:
-    logger.warning("Unable to find auto directory environment variable.")
+    general_logger.warning("Unable to find auto directory environment variable.")
 
 from auto2.continuations.fixed_points import FixedPointContinuation
 from auto2.continuations.periodic_orbits import PeriodicOrbitContinuation
@@ -56,7 +49,11 @@ class BifurcationDiagram(object):
         The name of the model to load. Used to load the .f90 file of the provided model.
     path_name: str, optional
         The directory path where files are read/saved.
+        If path is not absolute, it is relative to the current working directory.
         If `None`, defaults to current working directory.
+    logger: str or logging.Logger, optional
+        A logger to log the message of the bifurcation diagram.
+        Default to a given standard logger for each diagram.
 
     Attributes
     ----------
@@ -145,10 +142,9 @@ class BifurcationDiagram(object):
         Whether all the fixed points continuations possibilities of the bifurcation diagram have been computed.
     po_computed: bool
         Whether all the periodic orbits continuations possibilities of the bifurcation diagram have been computed.
-
     """
 
-    def __init__(self, model_name=None, path_name=None):
+    def __init__(self, model_name=None, path_name=None, logger=None):
 
         self.initial_points = None
         self.model_name = model_name
@@ -166,6 +162,35 @@ class BifurcationDiagram(object):
             self.config_object = ConfigParser(filepath)
         else:
             self.config_object = None
+
+        if isinstance(logger, str):
+            self._logger = logging.getLogger(logger)
+        elif isinstance(logger, logging.Logger):
+            self._logger = logger
+        elif logger is None:
+            if self.path_name is None:
+                path = './'
+            else:
+                path = self.path_name
+            log_string = f"{os.path.basename(os.path.abspath(path))}_{model_name}"
+            self._logger = logging.getLogger(f"{log_string}_logger")
+            self._logger.setLevel(logging.DEBUG)
+
+            formatter = logging.Formatter(
+                "%(asctime)s %(levelname)s: Model " + log_string + " -- Module %(filename)s -- %(message)s"
+            )
+
+            fh = logging.FileHandler(os.path.join(path, f"auto2_{model_name}.log"), mode="w", encoding="utf-8")
+            fh.setLevel(logging.DEBUG)
+            fh.setFormatter(formatter)
+            self._logger.addHandler(fh)
+
+            ch = logging.StreamHandler()
+            ch.setLevel(logging.INFO)
+            ch.setFormatter(formatter)
+            self._logger.addHandler(ch)
+        else:
+            raise ValueError('Provided logger not recognized.')
 
         self.fp_branches = dict()
         self.po_branches = dict()
@@ -208,6 +233,7 @@ class BifurcationDiagram(object):
         ----------
         path_name: str
             The path.
+            If path is not absolute, it is relative to the current working directory.
         """
         if os.path.exists(path_name):
             self._path_name = path_name
[…]
