### 2025-08-18 f6dd220
Finished bifurcation diagram restart and modified lrz demo to show usage
