+++ auto2/diagrams/bifurcations.py
@@ -9,6 +9,7 @@ This module defines the bifurcation diagram object used in auto-AUTO.
 
 import os
 import sys
+import shutil
 import warnings
 import logging
 import pickle
@@ -165,30 +166,15 @@ class BifurcationDiagram(object):
 
         if isinstance(logger, str):
             self._logger = logging.getLogger(logger)
+            self._log_string = False
         elif isinstance(logger, logging.Logger):
             self._logger = logger
+            self._log_string = False
         elif logger is None:
-            if self.path_name is None:
-                path = './'
-            else:
-                path = self.path_name
-            log_string = f"{os.path.basename(os.path.abspath(path))}_{model_name}"
-            self._logger = logging.getLogger(f"{log_string}_logger")
+            self._logger_string = f"{os.path.basename(os.path.abspath(self.path_name))}_{model_name}"
+            self._log_string = None
+            self._logger = logging.getLogger(f"{self._logger_string}_logger")
             self._logger.setLevel(logging.DEBUG)
-
-            formatter = logging.Formatter(
-                "%(asctime)s %(levelname)s: Model " + log_string + " -- Module %(filename)s -- %(message)s"
-            )
-
-            fh = logging.FileHandler(os.path.join(path, f"auto2_{model_name}.log"), mode="w", encoding="utf-8")
-            fh.setLevel(logging.DEBUG)
-            fh.setFormatter(formatter)
-            self._logger.addHandler(fh)
-
-            ch = logging.StreamHandler()
-            ch.setLevel(logging.INFO)
-            ch.setFormatter(formatter)
-            self._logger.addHandler(ch)
         else:
             raise ValueError('Provided logger not recognized.')
 
@@ -224,7 +210,18 @@ class BifurcationDiagram(object):
     @property
     def path_name(self):
         """str: The path where the |AUTO| and AUTO² files must be or are stored."""
-        return self._path_name
+        if self._path_name is not None:
+            return self._path_name
+        else:
+            return self.absolute_path_name
+
+    @property
+    def absolute_path_name(self):
+        """str: The absolute path where the |AUTO| and AUTO² files must be or are stored."""
+        if self._path_name is not None:
+            return os.path.abspath(self._path_name)
+        else:
+            return os.path.abspath('./')
 
     def set_path_name(self, path_name):
         """Set the path where the |AUTO| and AUTO² files must be or are stored.
@@ -242,6 +239,26 @@ class BifurcationDiagram(object):
                 "Path name given does not exist. Using the current working directory."
             )
             self._path_name = None
+
+    def _define_log_format(self, continuation_kwargs, mode='w', encoding="utf-8"):
+        if 'ICP' in continuation_kwargs:
+            cont_param = str(continuation_kwargs['ICP'][0])
+        else:
+            cont_param = str(self.config_object.continuation_parameters[0])
+        self._log_string = f'{self.model_name} -- Cont. Param. {cont_param}'
+        formatter = logging.Formatter(
+            "%(asctime)s %(levelname)s: Model " + self._log_string + " -- Module %(filename)s -- %(message)s"
+        )
+
+        fh = logging.FileHandler(os.path.join(self.path_name, f"auto2_{self.model_name}_{cont_param}.log"), mode=mode, encoding=encoding)
+        fh.setLevel(logging.DEBUG)
+        fh.setFormatter(formatter)
+        self._logger.addHandler(fh)
+
+        ch = logging.StreamHandler()
+        ch.setLevel(logging.INFO)
+        ch.setFormatter(formatter)
+        self._logger.addHandler(ch)
 
     def compute_fixed_points_diagram(
         self,
@@ -284,6 +301,9 @@ class BifurcationDiagram(object):
[…]
