+++ auto2/diagrams/bifurcations.py
@@ -171,12 +171,14 @@ class BifurcationDiagram(object):
             self._logger = logger
             self._log_string = False
         elif logger is None:
-            self._logger_string = f"{os.path.basename(os.path.abspath(self.path_name))}_{model_name}"
+            self._logger_string = (
+                f"{os.path.basename(os.path.abspath(self.path_name))}_{model_name}"
+            )
             self._log_string = None
             self._logger = logging.getLogger(f"{self._logger_string}_logger")
             self._logger.setLevel(logging.DEBUG)
         else:
-            raise ValueError('Provided logger not recognized.')
+            raise ValueError("Provided logger not recognized.")
 
         self.fp_branches = dict()
         self.po_branches = dict()
@@ -221,7 +223,7 @@ class BifurcationDiagram(object):
         if self._path_name is not None:
             return os.path.abspath(self._path_name)
         else:
-            return os.path.abspath('./')
+            return os.path.abspath("./")
 
     def set_path_name(self, path_name):
         """Set the path where the |AUTO| and AUTO² files must be or are stored.
@@ -240,17 +242,23 @@ class BifurcationDiagram(object):
             )
             self._path_name = None
 
-    def _define_log_format(self, continuation_kwargs, mode='w', encoding="utf-8"):
-        if 'ICP' in continuation_kwargs:
-            cont_param = str(continuation_kwargs['ICP'][0])
+    def _define_log_format(self, continuation_kwargs, mode="w", encoding="utf-8"):
+        if "ICP" in continuation_kwargs:
+            cont_param = str(continuation_kwargs["ICP"][0])
         else:
             cont_param = str(self.config_object.continuation_parameters[0])
-        self._log_string = f'{self.model_name} -- Cont. Param. {cont_param}'
+        self._log_string = f"{self.model_name} -- Cont. Param. {cont_param}"
         formatter = logging.Formatter(
-            "%(asctime)s %(levelname)s: Model " + self._log_string + " -- Module %(filename)s -- %(message)s"
+            "%(asctime)s %(levelname)s: Model "
+            + self._log_string
+            + " -- Module %(filename)s -- %(message)s"
         )
 
-        fh = logging.FileHandler(os.path.join(self.path_name, f"auto2_{self.model_name}_{cont_param}.log"), mode=mode, encoding=encoding)
+        fh = logging.FileHandler(
+            os.path.join(self.path_name, f"auto2_{self.model_name}_{cont_param}.log"),
+            mode=mode,
+            encoding=encoding,
+        )
         fh.setLevel(logging.DEBUG)
         fh.setFormatter(formatter)
         self._logger.addHandler(fh)
@@ -321,7 +329,9 @@ class BifurcationDiagram(object):
 
         if "NMX" not in continuation_kwargs:
             continuation_kwargs["NMX"] = 9000
-            self._logger.info("NMX parameters was not set, so setting it to 9000 points.")
+            self._logger.info(
+                "NMX parameters was not set, so setting it to 9000 points."
+            )
 
         self._logger.info("Continuing provided fixed points.")
         br_num = 1
@@ -632,7 +642,9 @@ class BifurcationDiagram(object):
 
         if "NMX" not in continuation_kwargs:
             continuation_kwargs["NMX"] = 9000
-            self._logger.info("NMX parameters was not set, so setting it to 9000 points.")
+            self._logger.info(
+                "NMX parameters was not set, so setting it to 9000 points."
+            )
 
         br_num = max(self.fp_branches.keys()) + 1
         self.level_reached = 0
@@ -909,11 +921,15 @@ class BifurcationDiagram(object):
                 self._logger.info(
                     "Restarting the computation of the periodic orbits from detected branching and period doubling points."
                 )
[…]
