### 2025-08-18 697a435
Implementing the restart of bifurcation diagrams along new continuation parameter (#43)
with lrz demo additions to show usage
