+++ auto2/diagrams/bifurcations.py
@@ -9,6 +9,7 @@ This module defines the bifurcation diagram object used in auto-AUTO.
 
 import os
 import sys
+import shutil
 import warnings
 import logging
 import pickle
@@ -21,14 +22,7 @@ from matplotlib.colors import TABLEAU_COLORS
 
 from auto2.parsers.config import ConfigParser
 
-logger = logging.getLogger("logger")
-logger.info(
-    "Using auto-AUTO (AUTO² or auto2) -- An AUTO-07p automatic search algorithm codebase"
-)
-logger.info(
-    "Read AUTO-07p manual first before using it. Wishing you a happy continuation, have fun !"
-)
-logger.info('Logging messages can be found in the file "auto2.log"')
+general_logger = logging.getLogger("general_logger")
 
 try:
     auto_directory = os.environ["AUTO_DIR"]
@@ -40,7 +34,7 @@ try:
         # sys.path.append(auto_directory + '/python/auto')
         sys.path.append(os.path.join(auto_directory, "python"))
 except KeyError:
-    logger.warning("Unable to find auto directory environment variable.")
+    general_logger.warning("Unable to find auto directory environment variable.")
 
 from auto2.continuations.fixed_points import FixedPointContinuation
 from auto2.continuations.periodic_orbits import PeriodicOrbitContinuation
@@ -56,7 +50,11 @@ class BifurcationDiagram(object):
         The name of the model to load. Used to load the .f90 file of the provided model.
     path_name: str, optional
         The directory path where files are read/saved.
+        If path is not absolute, it is relative to the current working directory.
         If `None`, defaults to current working directory.
+    logger: str or logging.Logger, optional
+        A logger to log the message of the bifurcation diagram.
+        Default to a given standard logger for each diagram.
 
     Attributes
     ----------
@@ -145,10 +143,9 @@ class BifurcationDiagram(object):
         Whether all the fixed points continuations possibilities of the bifurcation diagram have been computed.
     po_computed: bool
         Whether all the periodic orbits continuations possibilities of the bifurcation diagram have been computed.
-
     """
 
-    def __init__(self, model_name=None, path_name=None):
+    def __init__(self, model_name=None, path_name=None, logger=None):
 
         self.initial_points = None
         self.model_name = model_name
@@ -166,6 +163,22 @@ class BifurcationDiagram(object):
             self.config_object = ConfigParser(filepath)
         else:
             self.config_object = None
+
+        if isinstance(logger, str):
+            self._logger = logging.getLogger(logger)
+            self._log_string = False
+        elif isinstance(logger, logging.Logger):
+            self._logger = logger
+            self._log_string = False
+        elif logger is None:
+            self._logger_string = (
+                f"{os.path.basename(os.path.abspath(self.path_name))}_{model_name}"
+            )
+            self._log_string = None
+            self._logger = logging.getLogger(f"{self._logger_string}_logger")
+            self._logger.setLevel(logging.DEBUG)
+        else:
+            raise ValueError("Provided logger not recognized.")
 
         self.fp_branches = dict()
         self.po_branches = dict()
@@ -199,7 +212,18 @@ class BifurcationDiagram(object):
     @property
     def path_name(self):
         """str: The path where the |AUTO| and AUTO² files must be or are stored."""
-        return self._path_name
+        if self._path_name is not None:
+            return self._path_name
+        else:
+            return self.absolute_path_name
+
+    @property
+    def absolute_path_name(self):
+        """str: The absolute path where the |AUTO| and AUTO² files must be or are stored."""
+        if self._path_name is not None:
+            return os.path.abspath(self._path_name)
+        else:
+            return os.path.abspath("./")
 
     def set_path_name(self, path_name):
[…]
