+++ documentation/source/files/user_guide.rst
@@ -2,5 +2,133 @@
 User Guide
 ----------
 
-Coming soon...
+The general way to use auto-AUTO (or AUTO²) can be decomposed as follow:
+
+1. Writing the AUTO Fortran and configuration files describing the model to be studied. For this, we refer the user to the |AUTO| manual, although we will provide here an example.
+2. Scripting the exploration of the model bifurcation diagrams along different parameters, and representing them on plots.
+
+The second part requires in general the user to have some prerequisite knowledge of dynamical system theory and bifurcation theory.
+Along with the |AUTO| manual, these two books might help the user to get acquainted with theses theories:
+
+* Strogatz, S. H. (2024). Nonlinear dynamics and chaos: with applications to physics, biology, chemistry, and engineering. Chapman and Hall/CRC.
+* Kuznetsov, Y. A. (1998). Elements of applied bifurcation theory. New York, NY: Springer New York.
+
+Nevertheless, the present User Guide will provide you with a step-by-step detailed example, starting with the |AUTO| setup.
+
+1. Setting up the AUTO model files
+==================================
+
+Running a bifurcation analysis of a model with AUTO requires two different configuration files:
+
+1. A Fortran (90) file containing the tendencies of the model, coded as a fortran subroutine. Other procedure might also need to be implemented, depending on the complexity of the problem.
+2. A :code:`c.` configuration file, written with a Python-like synthax, and specifying the AUTO parameters of the runs (not to be confused with the model parameters).
+
+In the following, we shall use the celebrated Lorenz 63 (L63) model as an example to details how auto-AUTO works:
+
+.. math::
+
+    \dot x & = \sigma (y-x) \\
+    \dot y & = x \rho - y - x z \\
+    \dot z & = x y - \beta z
+
+Let's start by defining the Fortran tendencies file. This file should include a subroutine :code:`FUNC`.
+Here is such a subroutine for the L63 model:
+
+.. code-block:: fortran
+
+    SUBROUTINE FUNC(NDIM,U,ICP,PAR,IJAC,F,DFDU,DFDP)
+
+    IMPLICIT NONE
+    INTEGER, INTENT(IN) :: NDIM, ICP(*), IJAC
+    DOUBLE PRECISION, INTENT(IN) :: U(NDIM), PAR(*)
+    DOUBLE PRECISION, INTENT(OUT) :: F(NDIM)
+    DOUBLE PRECISION, INTENT(INOUT) :: DFDU(NDIM,NDIM), DFDP(NDIM,*)
+
+    DOUBLE PRECISION x, y, z, rho, beta, sigma
+
+    x=U(1)
+    y=U(2)
+    z=U(3)
+
+    rho=PAR(1)
+    beta=PAR(2)
+    sigma=PAR(3)
+
+    F(1)= sigma * (y-x)
+    F(2)= rho*x - y - x*z
+    F(3)= x*y -  beta*z
+
+    END SUBROUTINE FUNC
+
+We see that this subroutine template has several input arguments, but actually only the :code:`U` and :code:`PAR` are used in the code.
+The other arguments just correspond to the subroutine :code:`FUNC` signature defined by AUTO, but are not used here.
+:code:`U` is the state vector of the model :math:`\vec U = (x,y,z)`, and :code:`PAR` is a parameter vector whose dimension is fixed by AUTO to 36, each component corresponding to a
+parameter of the model. Here for example :code:`PAR(1)` is :math:`\rho`, :code:`PAR(2)` is :math:`\beta`, and :code:`PAR(3)` is :math:`\sigma`.
+
[…]
