### 2025-08-20 9faf3ec
Fixed point section
