+++ documentation/source/files/user_guide.rst
@@ -62,7 +62,8 @@ Here is such a subroutine for the L63 model:
 
 We see that this subroutine template has several input arguments, but actually only the :code:`U` and :code:`PAR` are used in the code.
 The other arguments just correspond to the subroutine :code:`FUNC` signature defined by AUTO, but are not used here.
-:code:`U` is the state vector of the model :math:`\vec U = (x,y,z)`, and :code:`PAR` is a parameter vector whose dimension is fixed by AUTO to 36, each component corresponding to a
+:code:`U` is the state vector of the model :math:`\vec U = (x,y,z)`, containing the variables defining the *phase space* of the system.
+:code:`PAR` is a parameter vector whose dimension is fixed by AUTO to 36, each component corresponding to a
 parameter of the model. Here for example :code:`PAR(1)` is :math:`\rho`, :code:`PAR(2)` is :math:`\beta`, and :code:`PAR(3)` is :math:`\sigma`.
 
 In addition to :code:`FUNC`, other subroutines must be provided:
@@ -135,13 +136,175 @@ in the reference of the :meth:`.FixedPointContinuation.make_continuation` and :m
 ============================
 
 Once your AUTO files are finalized, you can turn to creating a Python script or a `Jupyter`_ notebook.
-Let's create a Jupyter notebook called :code:`lrz.ipynb`. In general, you will start this script by loading some modules:
+In general, you will start this script by loading some modules:
 
-.. image:: figures/import_numpy.png
+.. code-block:: python
+
+    import numpy as np
+    from numba import njit
+    from scipy.optimize import root
+    from scipy.integrate import solve_ivp
+
+that will be useful later, and then load the :class:`.BifurcationDiagram` class:
+
+.. code-block:: python
+
+    from auto2.diagrams.bifurcations import BifurcationDiagram
 
 
+2.1 Finding initial solutions of the model
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+Computing a bifurcation diagram requires continuing solutions along a given parameter, but one needs solutions to start from at a given value of the
+said parameter. One can start from periodic orbit solutions, but the usual way is to start from a list of fixed (stationary) points.
+
+In our present example, they are the solutions of the equations:
+
+.. math::
+
+    \dot x & = \sigma (y-x) = 0 \\
+    \dot y & = x \rho - y - x z = 0 \\
+    \dot z & = x y - \beta z = 0
+
+i.e. the points :math:`(x_i,y_i,z_i)` where the tendencies :math:`(\dot x, \dot y, \dot z)` are zero.
+In the case of the L63 model, one could easily compute these solutions analytically, but here, for
+the sake of generality, we shall compute them numerically. For that we first need to code the tendencies equations:
+
+.. code-block:: python
+
+    @njit
+    def lrz(X, rho, beta, sigma):
+        x = X[0]
+        y = X[1]
+        z = X[2]
+        F = np.zeros(3)
+        F[0] = sigma * (y - x)
+        F[1] = rho * x - y - x * z
+        F[2] = x * y - beta * z
+        return F
+
+where we are using `Numba`_ to accelerate our code.
+In the following, we are going to compute the bifurcation diagram along :math:`\rho`, with the other parameters
+set to :math:`\sigma=10` and :math:`\beta=8/3`.
+We are thus going first to find the fixed points for the values :math:`\rho=0`, :math:`\sigma=10` and :math:`\beta=8/3`
+using a `root-finding algorithm`_ provided by `Scipy`_:
+
+.. code-block:: python
+
+    # parameters dictionary
+    params = {
+        'rho': 0.,
+        'beta': 8./3,
+        'sigma': 10.,
+    }
+
[…]
