### 2025-08-20 Climdyn/auto-AUTO#30 @epspebble
I still do not have access to my auto-07p setup (away from my work computer), but I have browsed through the new notebooks and am more than satisfied by the amount of quality work you have done to elucidate the use of autoAUTO!
