### 2025-08-20 Climdyn/auto-AUTO#29 @epspebble
Which parameter was not really the issue but I am convinced that you have provided this much needed functionality to potential users. Thank you again for the nice work.
