### 2025-08-21 f3758aa
Finishing first version of the user guide
