### 2025-08-21 b4b404a
Userguide (#44)
Adding a user guide to the documentation
