### 2025-08-21 6627521
Black
