+++ LICENSE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2025 Jonathan Demaeyer and Oísin Hamilton
+Copyright (c) 2025 Jonathan Demaeyer and Oisín Hamilton
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
