### 2025-08-22 78f3c0f
Update of the manuscript with more alternatives and a description of the future developments
