+++ documentation/notes/paper.md
@@ -67,11 +67,35 @@ knowledge about tipping points [@wunderling2021].
 
 # Existing Alternatives
 
+Over time, and alongside the development of AUTO, many tools were introduced to help the users deal with the complexity 
+of the output data stream of AUTO. We can mention for 
+example [XPPAUT](https://sites.pitt.edu/~phase/bard/bardware/xpp/xpp.html) [@ermentrout2012] which provides a 
+frontend to AUTO, making it easy to continue the solutions to boundary value problems as well as 
+equilibria, fixed points and limit cycles [@ermentrout2007].
+
+Another more recent development is [PyCoBi](https://github.com/pyrates-neuroscience/PyCoBi) which 
+*provides a Python interface to Auto-07p [1], allowing for a more intuitive usage of Auto-07p commands within Python scripts.*
+[@gast2025]
+
+Other bifurcation analysis software must also be mentioned here:
+
 * BifurcationKit.jl [@veltz2020]
 * PyDSTool [@clewley2007]
 * Other useful tools can be found at [https://dsweb.siam.org/Software](https://dsweb.siam.org/Software)
 
+However, to the authors knowledge, these tools do not provide automatic continuation features relevant to the 
+problems raised in the [Statement of Need](#statement-of-need). PyCoBi proposes limited automatic generation for 
+codimension-2 bifurcations, a feature that is also planned in auto-AUTO future developments.
 
+# Future Developments
+
+Future developments plan for auto-AUTO includes:
+
+* Automatic continuation of regime diagrams, i.e. codimension-1 branch tracking and detection of codimension-2 bifurcations by AUTO.
+* Investigate the formalization and abstraction of automatic continuation using graph theory, leveraging the functionalities already 
+  available in auto-AUTO, and allowing for a more insightful construction of the diagrams.
+
+The latter development is an obvious way to pursue the development of auto-AUTO, but is out of the scope of the authors research.
 
 # Acknowledgments
 
@@ -79,6 +103,7 @@ This project has received funding from the European Union’s Horizon 2020 resea
 the Marie Sklodowska–Curie grant agreement no. 956170. 
 In addition, funding has been provided through the "Fédération Wallonie-Bruxelles" with the 
 instrument "Fonds Spéciaux de Recherche".
+The authors thank the three referees for their constructive review and their suggestions that helped improve the manuscript.
 
 
 
