+++ documentation/notes/auto2.bib
@@ -120,3 +120,49 @@ Year = {2016}
   HAL_ID = {hal-02902346},
   HAL_VERSION = {v1},
 }
+
+@inbook{ermentrout2012,
+  author="Ermentrout, Bard",
+  editor="Le Nov{\`e}re, N.",
+  title="XPPAUT",
+  bookTitle="Computational Systems Neurobiology",
+  year="2012",
+  publisher="Springer Netherlands",
+  address="Dordrecht",
+  pages="519--531",
+  isbn="978-94-007-3858-4",
+  doi="10.1007/978-94-007-3858-4_17",
+  url="https://doi.org/10.1007/978-94-007-3858-4_17"
+}
+@article{ermentrout2007,
+    author = {Ermentrout, B. },
+    title   = {{X}{P}{P}{A}{U}{T}},
+    year    = {2007},
+    journal = {Scholarpedia},
+    volume  = {2},
+    number  = {1},
+    pages   = {1399},
+    doi     = {10.4249/scholarpedia.1399},
+    note    = {revision \#136177}
+}
+
+@software{gast2025,
+  author       = {Richard Gast},
+  title        = {pyrates-neuroscience/PyCoBi: Dropped support for
+                   Python < 3.9
+                  },
+  month        = apr,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v0.8.10},
+  doi          = {10.5281/zenodo.15232900},
+  url          = {https://doi.org/10.5281/zenodo.15232900},
+  swhid        = {swh:1:dir:f0982e298420a538cc070e619b170711c2c1fe5a
+                   ;origin=https://doi.org/10.5281/zenodo.7217789;vis
+                   it=swh:1:snp:70c99bf5d272cde968d2841ab968fd4322d21
+                   09d;anchor=swh:1:rel:a6dba133a1f7ee7a4e383ec532d4f
+                   88a2dd22293;path=pyrates-neuroscience-
+                   PyCoBi-d094998
+                  },
+}
+
