### 2025-08-22 e5d747f
Added the RP1982 figure instead of the L63
