+++ documentation/notes/auto2.bib
@@ -166,3 +166,14 @@ Year = {2016}
                   },
 }
 
+@article{reinhold1982,
+    author = {Reinhold, Brian B. and Pierrehumbert, Raymond T.},
+    doi = {10.1175/1520-0493(1982)110<1105:DOWRQS>2.0.CO;2},
+    journal = {Monthly Weather Review},
+    number = {9},
+    pages = {1105--1145},
+    title = {Dynamics of {{Weather Regimes}}: {{Quasi-Stationary Waves}} and {{Blocking}}},
+    url = {https://journals.ametsoc.org/view/journals/mwre/110/9/1520-0493_1982_110_1105_dowrqs_2_0_co_2.xml},
+    volume = {110},
+    year = {1982}
+}
