+++ documentation/notes/auto2.bib
@@ -122,7 +122,7 @@ Year = {2016}
 }
 
 @inbook{ermentrout2012,
-  author="Ermentrout, Bard",
+  author = {Ermentrout, B. },
   editor="Le Nov{\`e}re, N.",
   title="XPPAUT",
   bookTitle="Computational Systems Neurobiology",
