### 2025-08-25 #8079 @jodemaey
# Final response to @lockwo

Previous discussions:

* Initial review: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2854756076
* First reply: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2871751911
* Response by @lockwo: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2877509302

## Response to last message

Thanks again for reviewing our contribution

> > I don't understand this, there is no raw data provided to the code. What is provided to AUTO (and therefore to auto-AUTO) are equations files. Please explain what raw data you are talking about.
>
> Yes my point was unclear (but thinking about it more, also not that important), all is good in this regard.

Thanks.

> > Indeed, the user guide is pending. I hope to have the time to do it before next September. As for the example notebooks in the docs, I did that for a previous project and my experience is that it makes the repo heavier, due to the possible archiving of the past notebooks version figures. This time I wanted to avoid this, but fair enough. I will investigate a solution for this.
>
> That's an understandable goal, I wouldn't want bloat in the repo. I was using sphinx recently and had some success with making a /examples/ in docs that is just a symlink to the examples main directory. Then with `nbsphinx` and `nbsphinx_link` (and setting to never execute the files) in the conf.py. But I am not knowledgeable enough of sphinx to know if this is universally applicable. I can make a more detailed explanation when I get back from traveling next week.

The user guide is now available, and to me, it answers the point of having example notebooks in the documentation as well. I have not found an easy way to incorporate the example notebooks in a *light* way. It would increase the size of the repo by a few Mbs and I prefer to avoid this.

> > This is indeed missing. My current idea of testing this was to pytest the notebooks. This could be done soon, during the present review.
> > We could make more atomic tests, but then it would require some work of identifying the dynamical systems which present corresponding features to be tested (like continuation branches merging, crossing or looping over themselves). I (or [@ushham](https://github.com/ushham)) can do that in the future, but right now we don't have the bandwidth.
>
> I understand these full feature atomic tests could be quite some difficulty, so I think the pytest of the notebooks is definitely good. That ensures everything is working as expected.

Pytesting the notebooks is now available, and is used in a github workflow to check the PRs.
Thanks again for this suggestion.

> > It is true that there is no real comparison with other continuation packages, because that would be tantamount to comparing AUTO and the others. auto-AUTO on the other hand, is a layer on top of AUTO, implementing some data management, automatisms and continuation rules, and also plotting features, all of this in notebooks (while the underlying AUTO software has its own Python CLI). As such, there is not exactly an equivalent to auto-AUTO, as it is not per se a continuation software.
>
> I agree with this point, that auto^2 stands somewhat apart from other packages, and the features are not 1 to 1 comparable. That being said, perhaps it is worth including brief remark to this effect in the comparison section? I haven't worked much with the packages listed there, so I leave this to your discretion.
>

Based on the recommendation of @bartoldeman and our own findings, we now compare with similar codebases in the manuscript. The whole section has actually been rewritten:

"""
Over time, and alongside the development of AUTO, many tools were introduced to help the users deal with the complexity
of the output data stream of AUTO. We can mention for
example [XPPAUT](https://sites.pitt.edu/~phase/bard/bardware/xpp/xpp.html) [@ermentrout2012] which provides a
frontend to AUTO, making it easy to continue the solutions to boundary value problems as well as
equilibria, fixed points and limit cycles [@ermentrout2007].

Another more recent development is [PyCoBi](https://github.com/pyrates-neuroscience/PyCoBi) which
*provides a Python interface to Auto-07p, allowing for a more intuitive usage of Auto-07p commands within Python scripts.*
[@gast2025]

Other bifurcation analysis software must also be mentioned here:

* BifurcationKit.jl [@veltz2020]
* PyDSTool [@clewley2007]
* Other useful tools can be found at [https://dsweb.siam.org/Software](https://dsweb.siam.org/Software)

However, to the authors knowledge, these tools do not provide automatic continuation features relevant to the
problems raised in the [Statement of Need](#statement-of-need). PyCoBi proposes limited automatic generation for
codimension-2 bifurcation diagrams (regime diagrams), a feature that is also planned in auto-AUTO future developments.
"""

> > My personal view on this is that I use formatter for operational codes only. However, I would consider this if the number of contributors increases dramatically, but otherwise not.
>
> Fair enough, I suppose my proclivity for black is more of a philosophical opinion. After all, PEP 8 is really a [guideline more than a rule](https://www.youtube.com/watch?v=k9ojK9Q_ARE) 😄

After some thinking about this, we decided to go for black formatting. It is now enforced and checked before accepting PRs.
