### 2025-08-25 #8079 @jodemaey
# Final response to @epspebble

Previous discussions:

* Initial review: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2877917230
* First reply: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2885960403
* Reply by @epspebble: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2895631672
* Reply by me on high-dimensional systems: https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2912883098

## Issues raised

* https://github.com/Climdyn/auto-AUTO/issues/30 :
  Closed by:
  - https://github.com/Climdyn/auto-AUTO/pull/39
  - https://github.com/Climdyn/auto-AUTO/pull/40
  - https://github.com/Climdyn/auto-AUTO/pull/41
* https://github.com/Climdyn/auto-AUTO/issues/29 :
  Closed by:
  - https://github.com/Climdyn/auto-AUTO/pull/43

## High-dimensional case

To settle the [request for information about high-dimensional cases](https://github.com/openjournals/joss-reviews/issues/8079#issuecomment-2912883098), we give now the user some directions about how to use auto-AUTO in high-dimenional case, an example with a 20-dimensional model from geophysics is now proposed: https://github.com/Climdyn/auto-AUTO/blob/main/notebooks/geophysics/RP1982/RP1982.ipynb .

We think that this concludes our response to this review.
