### 2025-08-26 #8079 @lockwo
Yes, this completes my checklist. Setting up the formatting and automatic testing is really good for the repo and I appreciate the amount of work put into the package. It has a thumbs up from me 👍 .

Quick q, on line 46 of the paper I see what I believe is a quote from another work as italics, should that be italics (is that the formatting standard) or have quote marks?
