### 2025-08-28 #8079 @epspebble
@jodemaey @danielskatz

Finally I have access to my desktop and I love the new additions that have been done, especially the RP1982 example! I could confirm that the runtime was long, but well worth it (it took less than 20 minutes on a Ryzen 16-core processor, around 13 minutes). That strongly underscores the key unique advantage of this package - automation is really needed in longer calculations like this.

I am itching to try it out with a semi-discretization of a system of PDEs - it also naturally leads to a large system of ODEs, but that is not a comment for this review per se.

I am satisfied with all the replies and my checklist is now complete.
