### 2025-09-19 #8079 @bartoldeman
All my concerns and questions have been addressed, and @epspebble has verified the additions and in particular the RP1982 example, so @danielskatz , please proceed to acceptance.
