### 2025-09-19 94cad78
Update paper.md
