### 2025-09-19 7f91b71
Update auto2.bib
