+++ documentation/notes/auto2.bib
@@ -95,7 +95,7 @@ Year = {2016}
     number = {56},
     pages = {2597},
     author = {Jonathan Demaeyer and Lesley De Cruz and Stéphane Vannitsem},
-    title = {qgs: A flexible Python framework of reduced-order multiscale climate models},
+    title = {qgs: A flexible {P}ython framework of reduced-order multiscale climate models},
     journal = {Journal of Open Source Software},
 }
 
@@ -149,7 +149,7 @@ Year = {2016}
 @software{gast2025,
   author       = {Richard Gast},
   title        = {pyrates-neuroscience/PyCoBi: Dropped support for
-                   Python < 3.9
+                   {P}ython < 3.9
                   },
   month        = apr,
   year         = 2025,
@@ -172,7 +172,7 @@ Year = {2016}
     journal = {Monthly Weather Review},
     number = {9},
     pages = {1105--1145},
-    title = {Dynamics of {{Weather Regimes}}: {{Quasi-Stationary Waves}} and {{Blocking}}},
+    title = {Dynamics of Weather Regimes: Quasi-Stationary Waves and Blocking},
     url = {https://journals.ametsoc.org/view/journals/mwre/110/9/1520-0493_1982_110_1105_dowrqs_2_0_co_2.xml},
     volume = {110},
     year = {1982}
