### 2025-09-19 #8079 @danielskatz
@jodemaey - while we wait for @bartoldeman to check off the final two items, I've proofread your paper and suggest the changes in https://github.com/Climdyn/auto-AUTO/pull/47  Please merge this, or let me know what you disagree with.
