### 2025-09-20 b3b083c
Update of the Hamilton et al 2025 ref
