+++ documentation/notes/auto2.bib
@@ -100,12 +100,16 @@ Year = {2016}
 }
 
 @article{hamilton2025,
-    doi = {10.48550/arXiv.2503.02808},
-    url = {https://doi.org/10.48550/arXiv.2503.02808},
-    author = {Ois{\'i}n Hamilton and Jonathan Demaeyer and Michel Crucifix and Stéphane Vannitsem},
-    title = {Using Unstable Periodic Orbits to Understand Blocking Behaviour in a Low Order Land-Atmosphere Model},
-    journal = {Submitted to Chaos},
+    author = {Hamilton, Oisín and Demaeyer, Jonathan and Crucifix, Michel and Vannitsem, Stéphane},
+    title = {Using unstable periodic orbits to understand blocking behavior in a low order land–atmosphere model},
+    journal = {Chaos: An Interdisciplinary Journal of Nonlinear Science},
+    volume = {35},
+    number = {8},
+    pages = {083126},
     year = {2025},
+    month = {08},
+    doi = {10.1063/5.0268852},
+    url = {https://doi.org/10.1063/5.0268852},
 }
 
 @misc{veltz2020,
