+++ documentation/notes/auto2.bib
@@ -1,11 +1,13 @@
-hu@article{armstrongmckay2022,
+@article{armstrongmckay2022,
     author = {Armstrong McKay, David I. and Staal, Arie and Abrams, Jesse F. and Winkelmann, Ricarda and Sakschewski, Boris and Loriani, Sina and Fetzer, Ingo and Cornell, Sarah E. and Rockstr{\"o}m, Johan and Lenton, Timothy M.},
     journal = {Science},
     number = {6611},
     pages = {eabn7950},
     title = {Exceeding 1.5{$^\circ$}{{C}} Global Warming Could Trigger Multiple Climate Tipping Points},
     volume = {377},
-    year = {2022}
+    year = {2022},
+    doi = {10.1126/science.abn7950},
+    url = {https://www.science.org/doi/abs/10.1126/science.abn7950},
 }
 
 @article{ashwin2012,
@@ -41,7 +43,9 @@ hu@article{armstrongmckay2022,
     pages = {1786--1793},
     title = {Tipping Elements in the {{Earth}}'s Climate System},
     volume = {105},
-    year = {2008}
+    year = {2008},
+    doi = {10.1073/pnas.0705414105},
+    url = {https://www.pnas.org/doi/abs/10.1073/pnas.0705414105},
 }
 
 @article{wunderling2021,
@@ -51,7 +55,9 @@ hu@article{armstrongmckay2022,
     pages = {601--619},
     title = {Interacting Tipping Elements Increase Risk of Climate Domino Effects under Global Warming},
     volume = {12},
-    year = {2021}
+    year = {2021},
+    doi = {10.5194/esd-12-601-2021},
+    url = {https://esd.copernicus.org/articles/12/601/2021/},
 }
 
 @article{Hunter2007,
