### 2025-09-22 #8079 @jodemaey
Hi @danielskatz ,

I'm a bit confused here, because there are already release of the software (for example we are now at v0.5.4), and these release are already automatically archived on Zenodo. What should I do then ? Release a new version v0.5.5, and update the metadata on Zenodo ? But this would not contain any new development. I have troubles to see the point of this.
