### 2025-09-22 #8079 @danielskatz
@jodemaey - Given your practice of already archiving on zenodo, if the current version v0.5.4 matches the software as released by JOSS, perhaps you can just update the metadata for that version on zenodo. We want to make it clear that there is a release associated with the JOSS publication.
