### 2025-09-22 #8079 @danielskatz
1 is fine. Another option is a new release called v.0.5.4joss or similar, then you go back to the normal pattern after that.  But it's up to you.
