### 2025-09-27 #8079 @jodemaey
Hi @danielskatz ,

I have read the proofs one last time and it's fine for me.
Who is supposed to tell the editorial bot to accept ?
